requirements.txt
celery==5.3.6
Django==5.2
python-dotenv==1.1.0
django-environ==0.12.0
stripe==12.0.0
Pillow==11.1.0
uuid==1.30
django-crispy-forms==2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-compressor==4.5.1
django-stripe-payments==2.0.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
requests==2.32.3
pydantic==2.11.2
gunicorn==23.0.0
nltk==3.8.1
openai==1.72.0
python-decouple==3.8
python-docx==1.1.2
pymysql==1.1.1
PyPDF2==3.0.1
redis==5.0.1
tiktoken==0.9.0
whitenoise==6.9.0

